병합 및 코드 포맷 정리
