--- src/components/Lnb_2.tsx
@@ -1,112 +1,0 @@
-import { useState } from "react";
-import { Outlet } from "react-router-dom";
-import Modal from "./Modal";
-import TextInput from "./TextInput";
-import Button from "./Button";
-import FileUploadBox from "./FileUploadBox";
-import { useCreateProject } from "../api/project";
-
-const Lnb_2 = () => {
-  const { mutate } = useCreateProject();
-  const [isOpen, setIsOpen] = useState(true);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [file, setFile] = useState<File | null>(null);
-  const [formData, setFormData] = useState({
-    title: "",
-    description: "",
-  });
-
-  const modalOpen = () => {
-    setIsModalOpen(true);
-    setFile(null);
-    setFormData({
-      title: "",
-      description: "",
-    });
-  };
-
-  const onSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    console.log("입력값:", formData);
-    console.log("파일", file);
-    mutate({ title: formData.title });
-    setIsModalOpen(false);
-  };
-
-  const onChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    e.preventDefault();
-    setFormData((prev) => ({
-      ...prev,
-      [e.target.name]: e.target.value,
-    }));
-  };
-
-  const toggleSidebar = () => setIsOpen(!isOpen);
-
-  return (
-    <div className="flex h-screen">
-      {/* Sidebar */}
-      <div
-        className={`bg-gray-800 text-white transition-all duration-300 ${
-          isOpen ? "w-64" : "w-0 overflow-hidden"
-        }`}
-      >
-        <div className="whitespace-nowrap">LNB_2</div>
-        <button onClick={modalOpen}>프로젝트 생성</button>
-        <Modal
-          width={560}
-          isOpen={isModalOpen}
-          onClose={() => setIsModalOpen(false)}
-        >
-          <div className="text-2xl font-bold">새 프로젝트 생성</div>
-          <form
-            onSubmit={onSubmit}
-            className="flex flex-col items-center w-full gap-3"
-          >
-            <FileUploadBox onFileSelect={(f) => setFile(f)} />
-            <TextInput
-              inputType="text"
-              label="프로젝트명"
-              name="title"
-              placeholder="프로젝트 제목을 입력해주세요."
-              value={formData.title}
-              onChange={onChange}
-            />
-            <TextInput
-              inputType="textarea"
-              label="설명"
-              name="description"
-              placeholder="프로젝트에 대해 설명해주세요."
-              value={formData.description}
-              onChange={onChange}
-            />
-            <div className="flex self-end gap-2">
-              <Button
-                buttonType="button"
-                label="취소"
-                colorType="secondary"
-                onClick={() => setIsModalOpen(false)}
-              />
-              <Button buttonType="submit" label="생성" colorType="primary" />
-            </div>
-          </form>
-        </Modal>
-      </div>
-
-      {/* Main Content */}
-      <div className="relative flex-1 p-6 bg-gray-100">
-        <button
-          onClick={toggleSidebar}
-          className="absolute z-10 px-4 py-2 bg-white rounded shadow top-4 left-4"
-        >
-          {isOpen ? "close" : "open"}
-        </button>
-        <Outlet />
-      </div>
-    </div>
-  );
-};
-
-export default Lnb_2;
