+++ src/components/common/SearchInput.tsx
@@ -2,16 +2,16 @@
 import { useState } from "react";
 
 interface SearchInputProps {
-  enterEvent: () => void
+  enterEvent: () => void;
 }
 
-const SearchInput: React.FC<SearchInputProps> = ({enterEvent}) => {
+const SearchInput: React.FC<SearchInputProps> = ({ enterEvent }) => {
   const [inputValue, setInputValue] = useState("");
 
   const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
     if (e.key === "Enter") {
       // Enter일 경우 여기서 원하는 동작 수행
-      enterEvent()
+      enterEvent();
     }
   };
 
@@ -23,24 +23,24 @@ const SearchInput: React.FC<SearchInputProps> = ({enterEvent}) => {
         onKeyDown={handleKeyDown}
         onChange={(e) => setInputValue(e.target.value)}
         placeholder="프로젝트를 검색해라"
-        className="flex-grow outline-none bg-transparent placeholder-gray-400 text-sm"
+        className="flex-grow text-sm placeholder-gray-400 bg-transparent outline-none"
       />
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          className="w-5 h-5"
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke="currentColor"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M21 21l-4.35-4.35M16.65 16.65A7.5 7.5 0 1110.5 3a7.5 7.5 0 016.15 13.65z"
-          />
-        </svg>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        className="w-5 h-5"
+        fill="none"
+        viewBox="0 0 24 24"
+        stroke="currentColor"
+      >
+        <path
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          strokeWidth={2}
+          d="M21 21l-4.35-4.35M16.65 16.65A7.5 7.5 0 1110.5 3a7.5 7.5 0 016.15 13.65z"
+        />
+      </svg>
     </div>
   );
-}
+};
 
-export default SearchInput
+export default SearchInput;
