+++ src/components/lnb/ProjectProfile.tsx
@@ -1,18 +1,18 @@
 interface ProjectProfileProps {
-  url: string | null
-  title: string
+  url: string | null;
+  title: string;
 }
 
-const ProjectProfile: React.FC<ProjectProfileProps> = ({url, title}) => {
+const ProjectProfile: React.FC<ProjectProfileProps> = ({ url, title }) => {
   return (
     <div className="flex items-center justify-center w-11 h-11 rounded-[6px] bg-amber-600 hover:bg-amber-700 overflow-hidden text-sm font-medium cursor-pointer">
       {url ? (
-        <img src={url} alt={title} className="w-full h-full object-cover" />
+        <img src={url} alt={title} className="object-cover w-full h-full" />
       ) : (
         <span>{title.charAt(0).toUpperCase()}</span>
       )}
     </div>
-  )
-}
+  );
+};
 
-export default ProjectProfile
+export default ProjectProfile;
