+++ src/pages/Home.tsx
@@ -1,11 +1,9 @@
-import React from 'react'
-
 function Home() {
   return (
     <div className="flex justify-center h-screen align-middle">
       <h1>Home</h1>
     </div>
-  )
+  );
 }
 
-export default Home
+export default Home;
